+++ docs/en/docs/release-notes.md
@@ -12,6 +12,7 @@ hide:
 - Mark `http_exception` in the `Exception` as `async` to avoid thread creation.
 - Replace `asyncio.iscoroutinefunction` with `inspect.iscoroutinefunction`.
 - Replace `asyncio` in `run_sync` with anyio.
+- Refactor `run_sync` from `lilya.compat`.
 
 ## 0.16.10